refactor(sidebar): decompose UniversalSidebar into NavProvider + pure pieces [SHR-182]
Break the 1,712-line UniversalSidebar into a thin composition shell plus a
cohesive src/components/sidebar/ package, and sever two couplings that
blocked a standalone shell.

- Extract diff-tree localStorage helpers into pure @/lib/diff-tree-storage;
  DiffFileTree re-exports them for back-compat; BoardCard/TaskDetail/sidebar
  import the pure module directly.
- NavProvider holds shared persisted nav state (rail collapsed +
  collapsedGroups + toggles + group-sync), reusing GROUP_COLLAPSE_PREFIX/
  isGroupCollapsed from sidebar-nav.ts (de-dups the drifted constant).
- Move presentational pieces near-verbatim into glyphs/nav-items/nav-rows/
  row-menu/task-group/chats-section/chat-row-menu/footer; StatusIcon kept
  as-is.
- UniversalSidebar.tsx becomes a re-export barrel; App.tsx + integration
  test imports unchanged. Shell consumes Modular-06 useTasks() in place of
  useTasksContext().

Tests: 3326 passed; typecheck + lint clean. Adds glyphs.test.tsx and
nav-context.test.tsx.
