+++ src/components/DiffFileTree.tsx
@@ -4,6 +4,17 @@ import { cn } from '@/lib/utils';
 import { DIFF_TREE_ACTIVE, DIFF_TREE_ROW } from '@/lib/design-tokens';
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
 import { Button } from '@/components/ui/button';
+import { ignoredGroupKey, loadExpandedState, saveExpandedState } from '@/lib/diff-tree-storage';
+
+// Re-export the diff-tree storage helpers for back-compat with existing
+// importers. The implementations now live in `@/lib/diff-tree-storage`.
+export {
+  ignoredGroupKey,
+  diffTreeExpandedKey,
+  loadExpandedState,
+  saveExpandedState,
+  clearDiffTreeExpandedState,
+} from '@/lib/diff-tree-storage';
 
 interface Props {
   files: DiffFileEntry[];
@@ -14,54 +25,6 @@ interface Props {
   reviewed?: Set<string>;
   onToggleReview?: (path: string) => void;
   onToggleReviewed?: (path: string, currentlyReviewed: boolean) => void;
-}
-
-export function ignoredGroupKey(taskId: string): string {
-  return `octomux:diff-ignored-open:${taskId}`;
-}
-
-export function diffTreeExpandedKey(taskId: string): string {
-  return `octomux:diff-tree-expanded:${taskId}`;
-}
-
-/**
- * Read the persisted folder open-state overrides for a task. Returns a map of
- * folder path → open/closed; absent folders fall back to the default behaviour.
- * Returns `{}` when nothing is stored, the value is malformed, or localStorage
- * is unavailable (SSR, privacy mode, etc.).
- */
-export function loadExpandedState(taskId: string): Record<string, boolean> {
-  try {
-    const raw = localStorage.getItem(diffTreeExpandedKey(taskId));
-    if (!raw) return {};
-    const parsed: unknown = JSON.parse(raw);
-    if (!parsed || typeof parsed !== 'object' || Array.isArray(parsed)) return {};
-    const out: Record<string, boolean> = {};
-    for (const [key, value] of Object.entries(parsed)) {
-      if (typeof value === 'boolean') out[key] = value;
-    }
-    return out;
-  } catch {
-    return {};
-  }
-}
-
-export function saveExpandedState(taskId: string, state: Record<string, boolean>): void {
-  try {
-    localStorage.setItem(diffTreeExpandedKey(taskId), JSON.stringify(state));
-  } catch {
-    // ignore — persistence is best-effort
-  }
-}
-
-/** Best-effort cleanup of a task's persisted diff-tree state (e.g. on delete). */
-export function clearDiffTreeExpandedState(taskId: string): void {
-  try {
-    localStorage.removeItem(diffTreeExpandedKey(taskId));
-    localStorage.removeItem(ignoredGroupKey(taskId));
-  } catch {
-    // ignore — stale keys are harmless
-  }
 }
 
 interface Node {
