+++ src/components/sidebar/constants.ts
@@ -1,0 +1,12 @@
+// Shared layout + colour constants for the sidebar pieces. Kept verbatim from
+// the original monolithic UniversalSidebar so DOM/visuals are unchanged.
+
+export const FOCUS_RING = 'focus-ring';
+
+export const CYAN_ACTIVE_FG = '#7DD3FC';
+export const NAV_INACTIVE_FG = 'rgba(255,255,255,0.65)';
+export const ACTIVE_FILL = 'rgba(59,130,246,0.14)';
+export const ACTIVE_ACCENT = '#3B82F6';
+export const SIDEBAR_EXPANDED_WIDTH = 240;
+export const SIDEBAR_COLLAPSED_WIDTH = 56;
+export const RAIL_TILE_SIZE = 36;
